requirements.txt
asgiref==3.5.2
boto3==1.24.12
botocore==1.27.12
Django==4.0.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-rest-framework==0.1.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.1.1
psycopg2-binary==2.9.3
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.2.0
